chore: extracted legacy params code into separate file
